requirements.txt
streamlit
pandas
sqlalchemy
matplotlib
seaborn
google-generativeai
numpy
google-cloud
vertex-ai
